Refactored models module to several sub-modules in order to address the growing file-size.
